feat(webpack-transpiler): Import the webpack transpiler
* Rename `stryker-webpack` => `stryker-webpack-transpiler`.
* Make unit tests work on windows
* Add vscode project config
* Update stryker config to run in travis build
