+++ packages/stryker-webpack-transpiler/.editorconfig
