+++ packages/stryker-webpack-transpiler/.gitignore
