+++ packages/stryker-webpack-transpiler/.npmignore
