+++ packages/stryker-webpack-transpiler/.travis.yml
