+++ packages/stryker-webpack-transpiler/.vscode/settings.json
