+++ packages/stryker-webpack-transpiler/LICENSE
