+++ packages/stryker-webpack-transpiler/src/WebpackTranspiler.ts
