+++ packages/stryker-webpack-transpiler/src/compiler/Webpack.ts
@@ -1,0 +1,2 @@
+import * as webpack from 'webpack';
+export default webpack;
