+++ packages/stryker-webpack-transpiler/src/index.ts
