+++ packages/stryker-webpack-transpiler/stryker.conf.js
@@ -1,4 +1,4 @@
-module.exports = function(config) {
+module.exports = function (config) {
   config.set({
     files: [
       '!src/**/*.ts',
@@ -15,6 +15,23 @@ module.exports = function(config) {
     reporter: ["clear-text", "progress", "html"],
     tsconfigFile: "tsconfig.json",
     coverageAnalysis: "off",
-    logLevel: "info"
+    logLevel: "info",
+    thresholds: {
+      high: 90,
+      low: 80,
+      break: 80
+    },
+    plugins: [
+      '../../../stryker-html-reporter/src/index',
+      '../../../stryker-mocha-runner/src/index',
+      '../../../stryker-mocha-framework/src/index',
+      '../../../stryker-typescript/src/index'
+    ]
   });
+
+  if (process.env.TRAVIS) {
+    config.set({
+      maxConcurrentTestRunners: 2
+    });
+  }
 };
