+++ packages/stryker-webpack-transpiler/test/helpers/chaiPlugins.ts
@@ -1,0 +1,3 @@
+import * as sinonChai from 'sinon-chai';
+import * as chai from 'chai';
+chai.use(sinonChai);
