+++ packages/stryker-webpack-transpiler/test/helpers/initSourceMaps.ts
@@ -1,0 +1,1 @@
+import 'source-map-support/register';
