+++ packages/stryker-webpack-transpiler/test/helpers/mockInterfaces.ts
