+++ packages/stryker-webpack-transpiler/tslint.json
