--- packages/stryker-webpack/.vscode/launch.json
@@ -1,21 +1,0 @@
-{
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "node",
-            "request": "launch",
-            "name": "Launch Program",
-            "program": "${workspaceFolder}/node_modules/stryker/bin/stryker",
-            "args": [
-                "run"
-            ],
-            "outFiles": [
-                "${workspaceFolder}/**/*.js"
-            ],
-            "sourceMaps": true
-        }
-    ]
-}
