--- packages/stryker-webpack/src/compiler/Webpack.ts
@@ -1,2 +1,0 @@
-import * as webpack from "webpack";
-export default webpack;
